+++ Business/Data/Movie.cs
@@ -2,11 +2,12 @@ namespace PlumMediaCenter.Data
 {
     public class Movie
     {
-        public decimal? Id;
+        public ulong? Id;
         public string Title;
         public string Summary;
         public string Description;
         public string VideoPath;
         public string FolderPath;
+       
     }
 }
